Started on the comet enhancements
Removed the old in-ram persistence and moved it into a new object EntrySever which is a ListenerManager. The single entry page is now a CometActor which registerers with the EntrySever. Now I just need to implement the updates and clean up a bit and it should be ready for the screencast